AGENT/requirements.txt
asyncpg
python-dotenv
fastapi
langchain-google-vertexai==2.0.27
pandas
langgraph
langchain
uvicorn
langchain-mcp-adapters==0.1.8
matplotlib
